+++ src/main/resources/static/css/auth.css
@@ -85,3 +85,4 @@
     outline: none;
     border-bottom: 2px solid black;
 }
+
