+++ src/main/resources/static/css/listForm.css
@@ -1,0 +1,15 @@
+.postingCheckBox {
+    font-size: 18px;
+    background: #fff;
+    margin-left: 15px;
+    color: #0056b3;
+    border: none;
+}
+
+
+.postingCheckBox a:hover {
+    font-size: 18px;
+    box-shadow: inset 0 -10px 1px rgb(255, 85, 68);
+    color: #0056b3;
+    text-decoration: none;
+}
