+++ src/main/resources/static/css/styles.css
@@ -7,6 +7,7 @@
 @import url("footer.css");
 @import url("header.css");
 @import url("userjoin.css");
+@import url("listForm.css");
 @import url("text.css");
 @import url("post-list.css");
 @import url("detail.css");
